requirements.txt
torch
transformers
tqdm
pandas
scikit-learn
Pillow
openai
google-genai
pyyaml
